restructure components and re-organize files
