+++ src/components/smartSearch/SmartSearchDialog/FilterGallery.tsx
